requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fake-useragent==2.2.0
frozenlist==1.6.2
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.0.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.10.0
magic-filter==1.0.12
multidict==6.4.4
numpy==2.2.6
oauthlib==3.2.2
openai==1.84.0
openpyxl==3.1.5
orjson==3.10.18
outcome==1.3.0.post0
pandas==2.3.0
playwright==1.43.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyee==11.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
tenacity==9.1.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
uritemplate==4.2.0
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
